requirements.txt
PyGithub==2.3.0
python-dotenv==1.0.1
urllib3==2.2.2
openai==1.35.13
